catalog-backend: add type parameter to CatalogPermissionRule type
Previously the CatalogPermissionRule type had a fixed type of unknown[]
for the parameters expected in the `apply` and `toQuery` methods. This
meant that conditions generated for these rules would always have
unknown parameters too, which makes using them in policies much more
difficult.

To address this, this commit introduces a mandatory type parameter for
CatalogPermissionRule which is expected to be set to a tuple
corresponding to the expected parameters.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>